backend/pyproject.toml
[tool.poetry]
name = "imei-checker"
version = "0.1.0"
description = ""
authors = ["aleksandr.rybin <aleksandr.rybin@glowbyteconsulting.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.37"
fastapi = "^0.115.7"
alembic = "^1.14.1"
pydantic = "^2.10.5"
pydantic-settings = "^2.7.1"
asyncpg = "^0.30.0"
pyjwt = "^2.10.1"
uvicorn = "^0.34.0"
aiohttp = "^3.11.11"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
pytest-aiohttp = "^1.1.0"
loguru = "^0.7.3"
aiosqlite = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "session"
asyncio_mode = "auto"